% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc2ndBioDem.R
\name{calc2ndBioDem}
\alias{calc2ndBioDem}
\title{calc2ndBioDem}
\usage{
calc2ndBioDem(datasource, rev = numeric_version("0.1"))
}
\arguments{
\item{datasource}{source to be used}

\item{rev}{data revision the output will be produced for (numeric_version)}
}
\value{
magpie object with results on country level, weight on country level, unit and description.
}
\description{
calculates 2nd generation bioenergy demand
}
\examples{
\dontrun{
calcOutput("2ndBioDem")
}
}
