% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAtmosphericDepositionRates.R
\name{calcAtmosphericDepositionRates}
\alias{calcAtmosphericDepositionRates}
\title{calcAtmosphericDepositionRates}
\usage{
calcAtmosphericDepositionRates(cellular = FALSE)
}
\arguments{
\item{cellular}{TRUE for results on 0.5 degree grid.}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Conputes Atmospheric (nitrogen) deposition rates per area on different land-use types.
}
\examples{

\dontrun{ 
calcOutput("AtmosphericDepositionRates")
}
}
\seealso{
\code{\link{calcAtmosphericDeposition}},
\code{\link{calcNitrogenBudgetCropland}}
}
\author{
Benjamin Leon Bodirsky
}
