% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBrooks2005OldConservationPrios.R
\name{calcBrooks2005OldConservationPrios}
\alias{calcBrooks2005OldConservationPrios}
\title{calcBrooks2005OldConservationPrios}
\usage{
calcBrooks2005OldConservationPrios(cells = "lpjcell", nclasses = "seven")
}
\arguments{
\item{cells}{(deprecated) number of cells of landmask ("lpjcell" for 67420 cells)}

\item{nclasses}{Options are either "seven" or "nine".
\itemize{
\item "seven" separates primary and secondary forest and includes "crop", "past", "forestry",
"primforest", "secdforest", "urban" and "other"
\item "nine" adds the separation of pasture and rangelands, as well as a differentiation of
primary and secondary non-forest vegetation and therefore returns "crop", "past", "range",
"forestry", "primforest", "secdforest", "urban", "primother" and "secdother"
}}
}
\value{
magpie object in cellular resolution with different protection options in conservation priority areas
}
\description{
Function calculates land area in conservation priority areas
}
\examples{
\dontrun{
calcOutput("ConservationPriority", aggregate = FALSE)
}

}
\author{
Patrick v. Jeetze
}
