% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcClimateRegionsIPCC.R
\name{calcClimateRegionsIPCC}
\alias{calcClimateRegionsIPCC}
\title{calcClimateRegionsIPCC}
\usage{
calcClimateRegionsIPCC(
  landusetypes = "all",
  cellular = FALSE,
  yearly = FALSE,
  convert = TRUE
)
}
\arguments{
\item{landusetypes}{all or only one}

\item{cellular}{FALSE for country level, TRUE for cells}

\item{yearly}{FALSE for normal magpie 5 year time spans, TRUE for yearly}

\item{convert}{fills missing countries for country level aggregation with warm temperate moist (mostly small island nations)}
}
\value{
Country or cellular magpie object with fraction of each climate region by country or cell
}
\description{
calculates IPCC Climate Regions (IPCC2006 ch.4.3) based on t, ppt, pet from LPJml. elevation dimension not included for tropical montane class
}
\examples{

\dontrun{ 
calcOutput("ClimateRegionsIPCC")
}
}
\seealso{
\code{\link{readLPJml_rev21}}
}
\author{
David Chen
}
