% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcConservationPriority.R
\name{calcConservationPriority}
\alias{calcConservationPriority}
\title{calcConservationPriority}
\usage{
calcConservationPriority(cells = "magpiecell", nclasses = "seven")
}
\arguments{
\item{cells}{number of cells of landmask (select "magpiecell" for 59199 cells or "lpjcell" for 67420 cells)}

\item{nclasses}{Options are either "seven" or "nine".
\itemize{
\item "seven" separates primary and secondary forest and includes "crop", "past", "forestry",
"primforest", "secdforest", "urban" and "other"
\item "nine" adds the separation of pasture and rangelands, as well as a differentiation of
primary and secondary non-forest vegetation and therefore returns "crop", "past", "range",
"forestry", "primforest", "secdforest", "urban", "primother" and "secdother"
}}
}
\value{
magpie object in cellular resolution with different protection options in conservation priority areas
}
\description{
Function calculates land area in conservation priority areas
}
\examples{
\dontrun{
calcOutput("ConservationPriority", aggregate = FALSE)
}

}
\author{
Patrick v. Jeetze
}
