% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDomesticBalanceflow.R
\name{calcDomesticBalanceflow}
\alias{calcDomesticBalanceflow}
\title{calcDomesticBalanceflow}
\usage{
calcDomesticBalanceflow()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculate the corrective balance flow that is needed
because global production doesn't equal global supply
}
\examples{
\dontrun{
calcOutput("DomesticBalanceflow")
}
}
\seealso{
[calcFAOmassbalance()]
}
\author{
Benjamin Leon Bodirsky
}
