% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDomSupply.R
\name{calcDomSupply}
\alias{calcDomSupply}
\title{calcTradeMargin}
\usage{
calcDomSupply(nutrient = "dm", extend = TRUE)
}
\arguments{
\item{nutrient}{"dm"}

\item{extend}{extend into magpie future years}
}
\value{
magpie object domestic supply
}
\description{
Calculate regional domestic supply for trade module
}
\details{
Simple function to get the domestic supply for the trade module
}
\examples{
\dontrun{
a <- calcDomSupply()
}

}
\author{
David Chen
}
