% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEATLancetTargets.R
\name{calcEATLancetTargets}
\alias{calcEATLancetTargets}
\title{calcEATLancetTargets}
\usage{
calcEATLancetTargets(attributes = "kcal/d")
}
\arguments{
\item{attributes}{Attributes of food commodities (available: kcal/d and g/d)}
}
\value{
List of magpie objects with results on country level, weight on country level,
        unit and description.
}
\description{
Calculates minimum and maximum targets for healthy food intake according to reference
recommendations proposed by the EAT-Lancet Commission on healthy diets from sustainable
food systems, specified for different MAgPIE commodities.
}
\examples{
\dontrun{
calcOutput("EATLancetTargets")
}
}
\seealso{
\code{\link[madrat]{calcOutput}}, \code{\link{readEATLancet}},
\code{\link{calcEATLancetDiets}}
}
\author{
Isabelle Weindl
}
