% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEndUseTimber.R
\name{calcEndUseTimber}
\alias{calcEndUseTimber}
\title{calcEndUseTimber}
\usage{
calcEndUseTimber()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Calculates the demand of timber from historical FAO data (including intermediate products).
}
\examples{

\dontrun{ 
calcOutput("EndUseTimber")
}
}
\seealso{
\code{\link[mrfaocore]{calcFAOmassbalance_pre}}
}
\author{
Abhijeet Mishra
}
