% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcExoTcDummy.R
\name{calcExoTcDummy}
\alias{calcExoTcDummy}
\title{calcExoTcDummy}
\usage{
calcExoTcDummy()
}
\value{
Dummy file for regional exogenous tau path
}
\description{
Dummy file for regional exogenous tau path
}
\seealso{
\code{\link[madrat]{readSource}}, \code{\link[madrat]{calcOutput}}
}
\author{
Florian Humpenoeder
}
