% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFAOLossesWaste.R
\name{calcFAOLossesWaste}
\alias{calcFAOLossesWaste}
\title{calcFAOLossesWaste}
\usage{
calcFAOLossesWaste(out_type = "waste")
}
\arguments{
\item{out_type}{waste: food-specific ratios based on FAO food waste shares
waste_edible:  food-specific ratios based on FAO food waste shares including conversion into edible matter}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Calculates the ratio between food supply at household level and food intake for different MAgPIE commodities
based on estimated/assumed FAO waste shares for each commodity group (optionally also including food
conversion factors into edible matter).
}
\examples{
\dontrun{
calcOutput("FAOLossesWaste")
}
}
\seealso{
\code{\link[madrat]{calcOutput}}, \code{\link{readFAOLossesWaste}},
\code{\link{calcEATLancetWaste}}
}
\author{
Isabelle Weindl
}
