% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcForestAreaInitialization.R
\name{calcForestAreaInitialization}
\alias{calcForestAreaInitialization}
\title{calcForestAreaInitialization}
\usage{
calcForestAreaInitialization()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Calculates the management factor(s) needed to upscale the yield of forest plantations as compared to natural
vegetation based on FAO data.
}
\examples{
\dontrun{
calcOutput("ForestAreaInitialization")
}
}
\seealso{
\code{\link[mrfaocore]{calcFAOmassbalance_pre}}
}
\author{
Abhijeet Mishra
}
