% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcForestDisturbances.R
\name{calcForestDisturbances}
\alias{calcForestDisturbances}
\title{calcForestDisturbances}
\usage{
calcForestDisturbances()
}
\value{
MAgPIE object with FRA 2020 forest disturbance shares
}
\description{
Calculates which share of forest land is lost due to forest disturbances (including insects, diseases, severe weather events and other causes)
}
\examples{

\dontrun{ 
calcOutput("ForestDisturbances",aggregate=FALSE)
}


}
\seealso{
\code{\link[mrfaocore]{readFRA2020}}
}
\author{
Abhijeet Mishra
}
