% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcForestFireShare.R
\name{calcForestFireShare}
\alias{calcForestFireShare}
\title{calcForestFireShare}
\usage{
calcForestFireShare()
}
\value{
MAgPIE object with FRA 2020 forest fire shares
}
\description{
Calculates which share of forest land is lost due to forest fires
}
\examples{

\dontrun{ 
calcOutput("ForestFireShare",aggregate=FALSE)
}


}
\seealso{
\code{\link[mrfaocore]{readFRA2020}}
}
\author{
Abhijeet Mishra
}
