% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcForestGrassi2023.R
\name{calcForestGrassi2023}
\alias{calcForestGrassi2023}
\title{calcForestGrassi2023}
\usage{
calcForestGrassi2023()
}
\value{
List with a magpie object of intact and non-intact forest area (Mha) on cellular level
}
\description{
Intact and non-intact forest area at 0.5 degree after Grassi et al. (2023). Intact forest
follows Potapov et al. (2017) for 2013, with the masks for Canada, Brazil and Russia updated from national
information. Grassi et al. use non-intact forest as a proxy of the managed forest that national greenhouse
gas inventories report.
}
\examples{
\dontrun{
calcOutput("ForestGrassi2023", aggregate = FALSE)
}
}
\seealso{
\code{\link{readForestGrassi2023}}
}
\author{
Florian Humpenöder
}
