% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcForestryProductionRatio.R
\name{calcForestryProductionRatio}
\alias{calcForestryProductionRatio}
\title{calcForestryProductionRatio}
\usage{
calcForestryProductionRatio()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Calculates the management factor(s) needed to upscale the yield of forest plantations as compared to natural vegetation based on FAO data.
}
\examples{

\dontrun{ 
calcOutput("ForestryProductionRatio")
}
}
\seealso{
\code{\link[mrfaocore]{calcFAOmassbalance_pre}}
}
\author{
Abhijeet Mishra
}
