% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGDPdeflator.R
\name{calcGDPdeflator}
\alias{calcGDPdeflator}
\title{calcGDPdeflator}
\usage{
calcGDPdeflator(yearFrom = 2017, yearTo = 2005, currency = "PPP")
}
\arguments{
\item{yearFrom}{year in "y2005" format}

\item{yearTo}{year in "y2005" format}

\item{currency}{"PPP" or "MER"}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates a iso-level deflator,
this is needed to run food demand and livestock regressions consistently
}
\author{
David Chen
}
