% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGlobalSafetyNet.R
\name{calcGlobalSafetyNet}
\alias{calcGlobalSafetyNet}
\title{calcGlobalSafetyNet}
\usage{
calcGlobalSafetyNet(maginput = TRUE, nclasses = "seven", cells = "lpjcell")
}
\arguments{
\item{maginput}{Whether data should be transformed (based on LUH3 data) to match land use types used in MAgPIE.}

\item{nclasses}{If \code{magpie_input = TRUE}. Options are either "seven" or "nine". Note that by default,
the protected area is reported for urban land and forestry is zero.
\itemize{
\item "seven" separates primary and secondary forest and includes
"crop", "past", "forestry", "primforest", "secdforest", "urban" and "other"
\item "nine" adds the separation of pasture and rangelands, as well as a
differentiation of primary and secondary non-forest vegetation and therefore returns
"crop", "past", "range", "forestry", "primforest", "secdforest", "urban", "primother" and "secdother"
}}

\item{cells}{(deprecated) always lpjcell (67420 cells)}
}
\value{
List with a magpie object
}
\description{
Returns unprotected land area (Mha) within the Global Safety Net (Dinerstein et al. 2020).
}
\examples{
\dontrun{
calcOutput("calcGlobalSafetyNet", aggregate = FALSE)
}

}
\seealso{
\code{\link{readDinerstein2020}}
}
\author{
Patrick v. Jeetze
}
