% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGTAPTrade.R
\name{calcGTAPTrade}
\alias{calcGTAPTrade}
\title{calcGTAPTrade}
\usage{
calcGTAPTrade(subtype = NULL, bilateral = FALSE)
}
\arguments{
\item{subtype}{GTAP version and subtype, separated by "_"
available versions are "GTAP7", "GTAP8", and "GTAP9"
GTAP sheets relevant for trade are "VIWS": Trade - Bilateral Imports at World Prices
       "VIMS": Trade - Bilateral Imports at Market Prices
       "VXWD": Trade - Bilateral Exports at World Prices
       "VXMD": Trade - Bilateral Exports at Market Prices
       "VDFM": Intermediates - Firms' Domestic Purchases at Market Prices
       "VIFM": Intermediates - Firms' Imports at Market Prices
       "VFM": Endowments - Firms' Purchases at Market Prices
       "VOA": Payment received by producers (fram gtate value)
       "VOM": Value of output at dometic market prices}

\item{bilateral}{whether bilateral trade data should be calculated}
}
\value{
Trade related data as an MAgPIE object
}
\description{
calculate trade data from GTAP dataset
}
\examples{
\dontrun{
x <- calcGTAP("GTAP7_VXMD")
}
}
\author{
Xiaoxi Wang, David M Chen
}
