% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcH08evapotranspiration.R
\name{calcH08evapotranspiration}
\alias{calcH08evapotranspiration}
\title{calcH08evapotranspiration}
\usage{
calcH08evapotranspiration(subtype = "H08:mri-esm2-0")
}
\arguments{
\item{subtype}{Switch between different inputs}
}
\value{
magpie object in cellular resolution
}
\description{
Calc evapotranspiration data for SSP cenarios in mm/month
}
\examples{
\dontrun{
calcOutput("H08evapotranspiration", subtype = "H08:mri-esm2-0")
}

}
\author{
Marcos Alves
}
