% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHousehold_balanceflow.R
\name{calcHousehold_balanceflow}
\alias{calcHousehold_balanceflow}
\title{calcHousehold_balanceflow}
\usage{
calcHousehold_balanceflow()
}
\value{
magpie object in cellular resolution
}
\description{
Balance flow to make country-specific values on nutrition
outcome consistent with global homogeneous nutrition values. In case
of GE and Nr includes processing losses, in the case of DM just balances
unhomogeneous products.
}
\examples{
\dontrun{
calcOutput("calcHousehold_balanceflow")
}
}
\seealso{
[calcFAOmassbalance()]
}
\author{
Benjamin Bodirsky
}
