% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIPLCLand.R
\name{calcIPLCLand}
\alias{calcIPLCLand}
\title{calcIPLCLand}
\usage{
calcIPLCLand(maginput = TRUE, nclasses = "seven", datasource = "LandMark")
}
\arguments{
\item{maginput}{Whether data should be transformed (based on LUH3 data) to match land use types used in MAgPIE.}

\item{nclasses}{If \code{magpie_input = TRUE}. Options are either "seven" or "nine". Note that by default,
the protected area is reported for urban land and forestry is zero.
\itemize{
\item "seven" separates primary and secondary forest and includes
"crop", "past", "forestry", "primforest", "secdforest", "urban" and "other"
\item "nine" adds the separation of pasture and rangelands, as well as a
differentiation of primary and secondary non-forest vegetation and therefore returns
"crop", "past", "range", "forestry", "primforest", "secdforest", "urban", "primother" and "secdother"
}}

\item{datasource}{Currently only \code{"LandMark"}}
}
\value{
List with a magpie object
}
\description{
Returns unprotected land area (Mha) covered by Indigenous Peoples' and Local Community Lands.
}
\examples{
\dontrun{
calcOutput("calcIPLCLand", aggregate = FALSE)
}

}
\seealso{
\code{\link{readLandMark}}
}
\author{
Patrick v. Jeetze
}
