% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIr2RfYieldRatio.R
\name{calcIr2RfYieldRatio}
\alias{calcIr2RfYieldRatio}
\title{calcIr2RfYieldRatio}
\usage{
calcIr2RfYieldRatio()
}
\value{
MAgPIE object of yields
}
\description{
Passes on the irrigated to rainfed yield ratio from AQUASTAT
}
\examples{
  \dontrun{
    calcOutput("Ir2RfYieldRatio")
  }

}
\seealso{
[readAQUASTAT()],
[convertAQUASTAT()]
}
\author{
Kristine Karstens
}
