% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrecoverableCarbonLand.R
\name{calcIrrecoverableCarbonLand}
\alias{calcIrrecoverableCarbonLand}
\title{calcIrrecoverableCarbonLand}
\usage{
calcIrrecoverableCarbonLand(cells = "magpiecell")
}
\arguments{
\item{cells}{magpiecell (59199 cells) or lpjcell (67420 cells)}
}
\value{
List with a magpie object
}
\description{
Returns land area (Mha) that contains 50 %, 75% and
100% of irrecoverable carbon as defined in Noon et al (2022).
}
\examples{
\dontrun{
calcOutput("calcIrrecoverableCarbonLand", aggregate = FALSE)
}

}
\seealso{
\code{\link{readNoon2022}}
}
\author{
Patrick v. Jeetze
}
