% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigationInvCosts.R
\name{calcIrrigationInvCosts}
\alias{calcIrrigationInvCosts}
\title{calcIrrigationInvCosts}
\usage{
calcIrrigationInvCosts()
}
\value{
MAgPIE object
}
\description{
This function calculates irrigation investment costs for each country until
             the year 2050. Values linearly converge towards the value of Germany (1995)
             by 2050.
}
\examples{
\dontrun{
calcOutput("IrrigationInvCosts")
}

}
\seealso{
\code{\link{calcOutput}}, \code{\link{readWBirrigation}},
\code{\link{convertWBirrigation}}
}
\author{
Nele Steinmetz, Felicitas Beier
}
