% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcKeyBiodiversityAreas.R
\name{calcKeyBiodiversityAreas}
\alias{calcKeyBiodiversityAreas}
\title{calcKeyBiodiversityAreas}
\usage{
calcKeyBiodiversityAreas(
  maginput = TRUE,
  unprotected = TRUE,
  nclasses = "seven",
  cells = "lpjcell"
)
}
\arguments{
\item{maginput}{Whether data should be transformed (based on luh3 data) to match land use types used in MAgPIE.}

\item{unprotected}{if TRUE only KBA land that is currently unprotected is returned}

\item{nclasses}{If \code{magpie_input = TRUE}. Options are either "seven" or "nine". Note that by default,
the protected area is reported for urban land and forestry is zero.
\itemize{
\item "seven" separates primary and secondary forest and includes
"crop", "past", "forestry", "primforest", "secdforest", "urban" and "other"
\item "nine" adds the separation of pasture and rangelands, as well as a
differentiation of primary and secondary non-forest vegetation and therefore returns
"crop", "past", "range", "forestry", "primforest", "secdforest", "urban", "primother" and "secdother"
}}

\item{cells}{magpiecell (59199 cells) or lpjcell (67420 cells)}
}
\value{
List with a magpie object
}
\description{
Returns unprotected land area (Mha) within Key Biodiversity Areas.
}
\examples{
\dontrun{
calcOutput("calcKeyBiodiversityAreas", aggregate = FALSE)
}

}
\seealso{
\code{\link{readKeyBiodiversityAreas}}
}
\author{
Patrick v. Jeetze
}
