% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLossShare.R
\name{calcLossShare}
\alias{calcLossShare}
\title{calcLossShare}
\usage{
calcLossShare()
}
\value{
List of magpie object with results and weight on country or cellular level, unit and description.
}
\description{
Calculates share of domestic supply wasted
}
\examples{
\dontrun{
calcOutput("LossShare")
}
}
\author{
Benjamin Leon Bodirsky
}
