% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLPJmLClimateInput.R
\name{calcLPJmLClimateInput}
\alias{calcLPJmLClimateInput}
\title{calcLPJmLClimateInput}
\usage{
calcLPJmLClimateInput(
  subtype = "ISIMIP3bv2:IPSL-CM6A-LR:ssp126:1850-2100:tas:annual_mean",
  smooth = 0,
  cells = "magpiecell"
)
}
\arguments{
\item{subtype}{type of climate data to collect,
consisting of data source, GDM, RCP, time period, variable and time resolution
separated by ":"}

\item{smooth}{set averaging value for smoothing trajectories}

\item{cells}{number of halfdegree grid cells to be returned.
Options: "magpiecell" (59199), "lpjcell" (67420)}
}
\value{
magpie object in cellular resolution
}
\description{
calculate LPJmL climate inputs
}
\examples{
\dontrun{
calcOutput("LPJmLClimateInput", subtype = "ISIMIP3b:IPSL-CM6A-LR:ssp126:1850-2100:tas:annual_mean")
}

}
\author{
Kristine Karstens, Felicitas Beier
}
