% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMulticroppingCells.R
\name{calcMulticroppingCells}
\alias{calcMulticroppingCells}
\title{calcMulticroppingCells}
\usage{
calcMulticroppingCells(selectyears, lpjml, climatetype, scenario)
}
\arguments{
\item{selectyears}{Years to be returned}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{climatetype}{Switch between different climate scenarios or
historical baseline "GSWP3-W5E5:historical"}

\item{scenario}{"actual:total": currently multicropped areas calculated from total harvested areas
                and total physical areas per cell from readLandInG
"actual:crop" (crop-specific), "actual:irrigation" (irrigation-specific),
"actual:irrig_crop" (crop- and irrigation-specific) "total"
"potential:endogenous": potentially multicropped areas given
                        temperature and productivity limits
"potential:exogenous": potentially multicropped areas given
                       GAEZ suitability classification}
}
\value{
magpie object in cellular resolution
}
\description{
Returns grid cells where multiple cropping takes place given the
             chosen scenario
}
\examples{
\dontrun{
calcOutput("MulticroppingCells", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
