% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMulticroppingIntensity.R
\name{calcMulticroppingIntensity}
\alias{calcMulticroppingIntensity}
\title{calcMulticroppingIntensity}
\usage{
calcMulticroppingIntensity(
  scenario,
  selectyears,
  lpjml,
  climatetype,
  sectoral = "lpj"
)
}
\arguments{
\item{scenario}{"total": currently multicropped areas calculated from total harvested areas
         and total physical areas per cell from readLanduseToolbox
"crop" (crop-specific), "irrigation" (irrigation-specific),
"irrig_crop" (crop- and irrigation-specific)}

\item{selectyears}{Years to be returned}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{climatetype}{Switch between different climate scenarios or
historical baseline "GSWP3-W5E5:historical"}

\item{sectoral}{"kcr" MAgPIE crop types, and "lpj" LPJmL crop types}
}
\value{
magpie object in cellular resolution
}
\description{
Returns cropping intensity according to Landuse Toolbox
             given the chosen scenario
}
\examples{
\dontrun{
calcOutput("MulticroppingIntensity", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
