% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMulticroppingYieldIncrease.R
\name{calcMulticroppingYieldIncrease}
\alias{calcMulticroppingYieldIncrease}
\title{calcMulticroppingYieldIncrease}
\usage{
calcMulticroppingYieldIncrease(
  selectyears,
  lpjml,
  climatetype,
  fallowFactor = 0.75,
  suitability = "endogenous",
  crops = NULL
)
}
\arguments{
\item{selectyears}{Years to be returned}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}

\item{fallowFactor}{Factor determining yield reduction in off season due to
fallow period between harvest of first (main) season and
sowing of second (off) season}

\item{suitability}{"endogenous": suitability for multiple cropping determined
              by rules based on grass and crop productivity
"exogenous": suitability for multiple cropping given by
             GAEZ data set}

\item{crops}{NULL: default crops,
proxy: proxy crops for LPJmL to MAgPIE mapping and treatment of perennials}
}
\value{
magpie object in cellular resolution
}
\description{
Calculates yield increase achieved through multiple cropping
             (as factor of off season to main season crop yield) under
             irrigated and rainfed conditions respectively.
             Optionally: return which grid cells are potentially suitable for
             multiple cropping activities under rainfed and irrigated conditions.
             Calculation is based on grassland gross primary production (GPP)
             in the growing period of the respective crop and annual grass GPP.
}
\examples{
\dontrun{
calcOutput("MulticroppingYieldIncrease", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
