% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNINDiets.R
\name{calcNINDiets}
\alias{calcNINDiets}
\title{calcNINDiets}
\usage{
calcNINDiets(attributes = c("wm", "kcal"), calib = TRUE, FAOcountr = FALSE)
}
\arguments{
\item{attributes}{attributes of different food commodities (available: kcal and wm).}

\item{calib}{if TRUE, total daily per capita intake for MAgPIE food commodities is calibrated to NIN Lancet total intake.}

\item{FAOcountr}{if TRUE, estimates for countries not covered in FAOSTAT are set to Zero.}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Calculates daily per capita intake for MAgPIE food commodities that are consistent with diet scenarios
developed by the NIN-Lancet Commission on healthy diets from sustainable food systems.
The unit is kcal/day per capita or wm/day per capita.
Mapping of intake from NIN Lancet to MAgPIE food commodities is done indivudually for the different available units.
}
\examples{
\dontrun{
calcOutput("NINDiets")
}
}
\seealso{
\code{\link[madrat]{calcOutput}}, \code{\link{readNIN}},
\code{\link{convertNIN}}
}
\author{
Isabelle Weindl
}
