% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNINTargets.R
\name{calcNINTargets}
\alias{calcNINTargets}
\title{calcNINTargets}
\usage{
calcNINTargets(attributes = "kcal/d")
}
\arguments{
\item{attributes}{Attributes of food commodities (available: kcal/d and g/d)}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Calculates minimum and maximum targets for healthy food intake according to reference
recommendations proposed by the NIN on healthy diets from sustainable
food systems, specified for different MAgPIE commodities.
}
\examples{
\dontrun{
calcOutput("NINTargets")
}
}
\seealso{
\code{\link{calcOutput}}, \code{\link{readNIN}},
\code{\link{calcNINDiets}}
}
\author{
Isabelle Weindl
}
