% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNINWaste.R
\name{calcNINWaste}
\alias{calcNINWaste}
\title{calcNINWaste}
\usage{
calcNINWaste(out_type = "ratio")
}
\arguments{
\item{out_type}{ratio: total food supply to total intake.
ratio_detailed_calib: calibrated food-specific estimates.
ratio_detailed: food-specific estimates based on FAO food waste shares
calib: factor for calibrating estimates based on FAO waste shares to food supply}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Calculates the ratio between food supply at household level and food intake for different MAgPIE commodities
accounting for food-specific estimates of baseline intake of quantification of NIN diets by the
NIN comission, as well as for FAO food waste shares.
}
\examples{
\dontrun{
calcOutput("NINWaste")
}
}
\seealso{
\code{\link{calcOutput}}, \code{\link{readNIN}},
\code{\link{calcNINDiets}}, \code{\link{convertNIN}}
}
\author{
Isabelle Weindl
}
