% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNitrogenFixationFreelivingRate.R
\name{calcNitrogenFixationFreelivingRate}
\alias{calcNitrogenFixationFreelivingRate}
\title{calcNitrogenFixationFreelivingRate}
\usage{
calcNitrogenFixationFreelivingRate()
}
\value{
List of magpie objects with results on global level, empty weight, unit and description.
}
\description{
calculates fixation rates from freeliving bacteria per area
}
\examples{

\dontrun{
calcOutput("NitrogenFixationFreelivingRate")
}

}
\seealso{
\code{\link[mrcommons]{readHerridge}}
}
\author{
Benjamin Leon Bodirsky
}
