% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNutritionAttributes.R
\name{calcNutritionAttributes}
\alias{calcNutritionAttributes}
\title{calcNutritionAttributes}
\usage{
calcNutritionAttributes()
}
\value{
magpie object
}
\description{
Calculates nutrition attributes of food products, i.e. calorie and protein
             supply of a product dedicated to food use.
}
\examples{
\dontrun{
calcOutput("NutritionAttributes", aggregate = FALSE)
}

}
\author{
Benjamin Bodirsky
}
