% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlantationContribution.R
\name{calcPlantationContribution}
\alias{calcPlantationContribution}
\title{calcPlantationContribution}
\usage{
calcPlantationContribution()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Calculates the interpolated contribution share of plantations to roundwood demand
}
\examples{
\dontrun{
calcOutput("PlantationContribution")
}
}
\seealso{
\code{\link[mrfaocore]{calcFAOmassbalance_pre}}
}
\author{
Abhijeet Mishra
}
