% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlantedForest.R
\name{calcPlantedForest}
\alias{calcPlantedForest}
\title{calcPlantedForest}
\usage{
calcPlantedForest()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Calculates the share of plantations in planted forest
}
\examples{
\dontrun{
calcOutput("PlantedForest")
}
}
\seealso{
\code{\link[mrfaocore]{calcFAOmassbalance_pre}}
}
\author{
Abhijeet Mishra
}
