% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPotentialEvapoTrans.R
\name{calcPotentialEvapoTrans}
\alias{calcPotentialEvapoTrans}
\title{calcPotentialEvapoTrans}
\usage{
calcPotentialEvapoTrans(
  climatetype = "MRI-ESM2-0:ssp370",
  datasource = "LPJmLClimateInput:LPJmL4_for_MAgPIE_44ac93de:harmonized2020"
)
}
\arguments{
\item{climatetype}{Switch between different climate scenarios}

\item{datasource}{Source of the data. Available at the moment: LPJmLClimateInput}
}
\value{
magpie object in cellular resolution
}
\description{
Calculates or loads potential Evapotranspiration
}
\examples{
\dontrun{
calcOutput("calcPotentialEvapoTrans", datasource = "LPJmLClimateInput")
}

}
\author{
Kristine Karstens
}
