% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcProcessing_balanceflow.R
\name{calcProcessing_balanceflow}
\alias{calcProcessing_balanceflow}
\title{calcProcessing_balanceflow}
\usage{
calcProcessing_balanceflow()
}
\value{
magpie object of secondary product balanceflow
}
\description{
Calculates the difference between
what is recorded in FAOSTAT production of secondary
products, and what is calculated from our global
processing conversion factors in order to meet FAOSTAT
production. Fades out by 2050.
}
\examples{
\dontrun{
a <- calcOutput("Processing_balanceflow",
                aggregate = FALSE)
}
}
\seealso{
\code{\link[mrland]{calcProcessing_conversion_factors}}
}
\author{
Benjamin Bodirsky, David M Chen
}
