% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcProcessing_conversion_factors.R
\name{calcProcessing_conversion_factors}
\alias{calcProcessing_conversion_factors}
\title{calcProcessing_conversion_factors}
\usage{
calcProcessing_conversion_factors()
}
\value{
magpie object of conversion factors
}
\description{
Calculates global conversion factors
from the FAOmassbalance_pre in order to feed to MAgPIE for
conversion of primary to secondary products
}
\examples{
\dontrun{
a <- calcOutput("Processing_conversion_factors",
                aggregate = FALSE)
}
}
\author{
Benjamin Bodirsky, David M Chen
}
