% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcProtectedAreaBaseline.R
\name{calcProtectedAreaBaseline}
\alias{calcProtectedAreaBaseline}
\title{calcProtectedAreaBaseline}
\usage{
calcProtectedAreaBaseline(cells = "magpiecell")
}
\arguments{
\item{cells}{magpiecell (59199 cells) or lpjcell (67420 cells)}
}
\value{
List with a magpie object
}
\description{
Returns protected land area (Mha) in terms of cropland, pasture, forest and other land between 1995 and 2020.
}
\examples{
\dontrun{
calcOutput("ProtectedAreaBaseline", aggregate = FALSE)
}

}
\seealso{
\code{\link{calcProtectArea}}
}
\author{
Patrick v. Jeetze
}
