% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPumpingCosts.R
\name{calcPumpingCosts}
\alias{calcPumpingCosts}
\title{calcPumpingCosts}
\usage{
calcPumpingCosts()
}
\value{
A magpie object at iso level for all years with information on pumping costs
}
\description{
provides costs of pumping irrigation water
}
\examples{
\dontrun{
calcOutput("PumpingCosts")
}

}
\author{
Vartika Singh
#' @seealso \code{\link[madrat]{readSource}}, \code{\link[madrat]{calcOutput}}
}
