% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSeedShare.R
\name{calcSeedShare}
\alias{calcSeedShare}
\title{calcSeedShare}
\usage{
calcSeedShare()
}
\value{
List of magpie object with results and weight on country or cellular level, unit and description.
}
\description{
Calculates Seed share (seed demand per production)
}
\examples{
\dontrun{
calcOutput("SeedShare")
}
}
\author{
Benjamin Leon Bodirsky
}
