% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSoilStockChangeFactors.R
\name{calcSoilStockChangeFactors}
\alias{calcSoilStockChangeFactors}
\title{calcSoilStockChangeFactors}
\usage{
calcSoilStockChangeFactors()
}
\value{
MAgPIE object of yields
}
\description{
calculates and merges information on stock change factors
}
\examples{
  \dontrun{
    calcOutput("SoilStockChangeFactors")
  }

}
\seealso{
[readIPCC()]
}
\author{
Kristine Karstens
}
