% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMexogenous.R
\name{calcSOMexogenous}
\alias{calcSOMexogenous}
\title{calcSOMexogenous}
\usage{
calcSOMexogenous()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Uses an exogenous trajectory of Soil organic matter loss nitrogen release
}
\examples{
\dontrun{
calcOutput("SOMexogenous")
}
}
\seealso{
\code{\link[madrat]{calcOutput}}
}
\author{
Benjamin Leon Bodirsky
}
