% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTauHistorical.R
\name{calcTauHistorical}
\alias{calcTauHistorical}
\title{calcTauHistorical}
\usage{
calcTauHistorical()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Calculates historical trends in agricultural land use intensity Tau based on FAO yield trends.
}
\examples{

\dontrun{ 
calcOutput("TauHistorical")
}
}
\author{
Isabelle Weindl
}
