% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTimberDemandExt.R
\name{calcTimberDemandExt}
\alias{calcTimberDemandExt}
\title{calcTimberDemandExt}
\usage{
calcTimberDemandExt()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Calculates the demand of timber from FAO data (including intermediate products).
}
\examples{

\dontrun{ 
calcOutput("TimberDemandExt")
}
}
\seealso{
\code{\link[mrfaocore]{calcFAOmassbalance_pre}}
}
\author{
Abhijeet Mishra
}
