% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTimberHarvestCost.R
\name{calcTimberHarvestCost}
\alias{calcTimberHarvestCost}
\title{calcTimberHarvestCost}
\usage{
calcTimberHarvestCost()
}
\value{
List of magpie object with results on country level, weight on country level, unit and description.
}
\description{
Calculates the cost for timber harvest
}
\examples{

\dontrun{
calcOutput("TimberHarvestCost")
}
}
\author{
Abhijeet Mishra
}
