% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTradeBalance.R
\name{calcTradeBalance}
\alias{calcTradeBalance}
\title{Calculate imports/exports}
\usage{
calcTradeBalance()
}
\value{
regional trade balances
}
\description{
Calculate the difference between production and
domestic_supply. Numbers till 2010 are derived from FAO.
Numbers after 2010 are hold constant
}
\examples{

\dontrun{ 
a <- calcTradeBalance()
}
}
\seealso{
\code{\link[madrat]{calcOutput}}, \code{\link[mrcommons]{calcFAOmassbalance}}
}
\author{
Jan Philipp Dietrich
}
