% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTradeBilateralFAOHarmonized.R
\name{calcTradeBilateralFAOHarmonized}
\alias{calcTradeBilateralFAOHarmonized}
\title{calcTradeBilateralFAOHarmonized}
\usage{
calcTradeBilateralFAOHarmonized(yearly = TRUE)
}
\arguments{
\item{yearly}{whether to calculate yearly data or only magpie 5year timesteps}
}
\value{
Self import to dupply ratio
}
\description{
Harmonizes bilateral trade values to FAO mass balance imports
often termed "Import Dependency Ratio"
}
\examples{
\dontrun{
a <- calcTradeBilateralFAOHarmonized()
}

}
\seealso{
\code{\link[madrat]{calcOutput}}, \code{\link[mrcommons]{calcFAOmassbalance}}
}
\author{
David M Chen
}
