% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTradeBilateralScenarioAdj.R
\name{calcTradeBilateralScenarioAdj}
\alias{calcTradeBilateralScenarioAdj}
\title{calcTradeBilateralScenarioAdj}
\usage{
calcTradeBilateralScenarioAdj()
}
\value{
dummy file with 0's to befilled in MAgPIE
}
\description{
Creates dummy trade scenario adjustments file filld with 0's,
actual changes currently implemented in the model on the import supply ratio
(imports/domestic supply)
}
\author{
David M Chen
}
