% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTradeExportShr.R
\name{calcTradeExportShr}
\alias{calcTradeExportShr}
\title{calcTradeExportShr}
\usage{
calcTradeExportShr()
}
\value{
export shares as MAgPIE object
}
\description{
Calculate  export shares of countries compared to total export. 
             This is based on export values from FAOSTAT. 
             Function calculates this based on average
             values of the specified years.
}
\examples{
\dontrun{
a <- calcTradeExportShr()
}

}
\seealso{
\code{\link[madrat]{calcOutput}}, \code{\link[mrcommons]{calcFAOmassbalance}}
}
\author{
Ulrich Kreidenweis, Xiaoxi Wang
}
