% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTradeImportSupplyRatio.R
\name{calcTradeImportSupplyRatio}
\alias{calcTradeImportSupplyRatio}
\title{calcTradeImportSupplyRatio}
\usage{
calcTradeImportSupplyRatio(magYears = FALSE)
}
\arguments{
\item{magYears}{whether to output in magpie 5year timesteps}
}
\value{
Self import to supply ratio
}
\description{
Calculates regional imports to supply ratios
often termed "Import Dependency Ratio"
}
\examples{
\dontrun{
a <- calcTradeImportSupplyRatio()
}

}
\seealso{
\code{\link[madrat]{calcOutput}}, \code{\link[mrcommons]{calcFAOmassbalance}}
}
\author{
David M Chen
}
