% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTradeSelfSuff.R
\name{calcTradeSelfSuff}
\alias{calcTradeSelfSuff}
\title{calcTradeSelfSuff}
\usage{
calcTradeSelfSuff()
}
\value{
Self sufficiences
}
\description{
Calculates regional self sufficiences from FAO data as
production/domestic_supply.
}
\examples{
\dontrun{
a <- calcTradeSelfSuff()
}

}
\seealso{
\code{\link[madrat]{calcOutput}}, \code{\link[mrcommons]{calcFAOmassbalance}}
}
\author{
Ulrich Kreidenweis
}
