% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValueProduction.R
\name{calcValueProduction}
\alias{calcValueProduction}
\title{calcValueProduction}
\usage{
calcValueProduction(datasource = "FAO", cellular = TRUE)
}
\arguments{
\item{datasource}{Options of the source of the price data: only FAO has country level data}

\item{cellular}{cellular or iso country values}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates production value based on production and prices, only works for FAO dataset currently
}
\examples{
\dontrun{
calcOutput("ValueProduction")
}

}
\seealso{
\code{\link[mrcommons]{calcProduction}},
\code{\link[mrcommons]{calcPriceAgriculture}}
}
\author{
David Chen
}
