% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcYields.R
\name{calcYields}
\alias{calcYields}
\title{calcYields}
\usage{
calcYields(
  source = c(lpjml = "ggcmi_phase3_nchecks_9ca735cb", isimip = NULL),
  climatetype = "GSWP3-W5E5:historical",
  cells = "magpiecell",
  weighting = "totalCrop",
  multicropping = FALSE,
  indiaYields = FALSE,
  scaleFactor = 0.3
)
}
\arguments{
\item{source}{Defines LPJmL version for main crop inputs and isimip replacement.
For isimip choose crop model/gcm/rcp/co2 combination formatted like this:
"yields:EPIC-IIASA:ukesm1-0-ll:ssp585:default:3b"}

\item{climatetype}{Switch between different climate scenarios}

\item{cells}{if cellular is TRUE: "magpiecell" for 59199 cells or "lpjcell" for 67420 cells}

\item{weighting}{use of different weights (totalCrop (default), totalLUspecific, cropSpecific, crop+irrigSpecific,
avlCropland, avlCropland+avlPasture)}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE) and
Multiple Cropping Suitability mask selected
(mask can be:
"none": no mask applied (only for development purposes)
"actual:total": currently multicropped areas calculated from total harvested areas
                and total physical areas per cell from readLanduseToolbox
"actual:crop" (crop-specific), "actual:irrigation" (irrigation-specific),
"actual:cropIrrig" (crop- and irrigation-specific) "total"
"potential:endogenous": potentially multicropped areas given
                        temperature and productivity limits
"potential:exogenous": potentially multicropped areas given
                       GAEZ suitability classification)
(e.g. TRUE:actual:total; TRUE:none; FALSE)}

\item{indiaYields}{if TRUE returns scaled yields for rainfed crops in India}

\item{scaleFactor}{integer value by which indiaYields will be scaled}
}
\value{
magpie object in cellular resolution
}
\description{
This function extracts yields from LPJmL
             and transforms them to MAgPIE crops calibrating proxy crops
             to FAO yields. Optionally, ISIMIP yields can be returned.
}
\examples{
\dontrun{
calcOutput("Yields", aggregate = FALSE)
}

}
\author{
Kristine Karstens, Felicitas Beier
}
