% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcYieldsLPJmL.R
\name{calcYieldsLPJmL}
\alias{calcYieldsLPJmL}
\title{calcYieldsLPJmL}
\usage{
calcYieldsLPJmL(
  lpjml = "ggcmi_phase3_nchecks_bft_e511ac58",
  climatetype = "GSWP3-W5E5:historical",
  cells = "lpjcell"
)
}
\arguments{
\item{lpjml}{Defines LPJmL version for main crop inputs}

\item{climatetype}{Switch between different climate scenarios}

\item{cells}{if cellular is TRUE: "magpiecell" for 59199 cells
or "lpjcell" for 67420 cells}
}
\value{
magpie object in cellular resolution
}
\description{
This function extracts yields from LPJmL
             for all years
}
\examples{
\dontrun{
calcOutput("YieldsLPJmL", aggregate = FALSE)
}

}
\author{
Kristine Karstens, Felicitas Beier
}
