% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcYldPastSwitch.R
\name{calcYldPastSwitch}
\alias{calcYldPastSwitch}
\title{calcYldPastSwitch}
\usage{
calcYldPastSwitch()
}
\value{
Dummy file for regional pasture yield switch
}
\description{
Dummy file for regional pasture yield switch (f14_yld_past_switch)
}
\seealso{
\code{\link[madrat]{readSource}}, \code{\link[madrat]{calcOutput}}
}
\author{
Alexandre Köberle, Patrick Rein
}
