% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertForestryProductionRatio.R
\name{convertForestryProductionRatio}
\alias{convertForestryProductionRatio}
\title{Converts Forestry Production Ratio
Update dd-Jmm-jjjj - Please add comment if changes made here (Abhi)}
\usage{
convertForestryProductionRatio(x)
}
\arguments{
\item{x}{MAgPIE object to be converted}
}
\value{
A MAgPIE object containing country disaggregated data
}
\description{
Converts Forestry Production Ratio
Update dd-Jmm-jjjj - Please add comment if changes made here (Abhi)
}
\examples{

\dontrun{ a <- readSource("ForestryProductionRatio",convert=FALSE)
}

}
\author{
Abhijeet Mishra
}
