% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGFWLossByDriver.R
\name{convertGFWLossByDriver}
\alias{convertGFWLossByDriver}
\title{convertGFWLossByDriver}
\usage{
convertGFWLossByDriver(x)
}
\arguments{
\item{x}{magpie object as returned by \code{\link{readGFWLossByDriver}}}
}
\value{
magpie object on madrat's ISO country set, Mha
}
\description{
Moves the GFW data from the GADM country set to madrat's ISO country set: Kosovo,
Northern Cyprus and the Akrotiri and Dhekelia base areas are added to Serbia and Cyprus, GADM's
placeholder codes for disputed territories are dropped and countries GFW does not list are
filled with zero.
}
\examples{
\dontrun{
a <- readSource("GFWLossByDriver", convert = TRUE)
}
}
\seealso{
\code{\link{readGFWLossByDriver}}
}
\author{
Michael Crawford
}
