% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertInocencio2007.R
\name{convertInocencio2007}
\alias{convertInocencio2007}
\title{convertInocencio2007}
\usage{
convertInocencio2007(x)
}
\arguments{
\item{x}{MAgPIE object containing irrigation cost data at region level resolution}
}
\value{
MAgPIE object on ISO country level
}
\description{
Convert irrigation cost data by Inocencio et al. (2007)
             from regional to ISO country resolution
}
\examples{
\dontrun{
a <- convertInocencio2005(x)
}

}
\author{
Felicitas Beier
}
