% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertNIN.R
\name{convertNIN}
\alias{convertNIN}
\title{Convert data from the NIN Lancet Comission}
\usage{
convertNIN(x, subtype)
}
\arguments{
\item{x}{MAgPIE object containing NIN Lancet data at mixed country-region
resolution}

\item{subtype}{Type of NIN Lancet data that should be read. Available types are:
\itemize{
\item \code{cons_data}: Consumption analysis ("NIN_Lancet_cons_data.csv")
\item \code{recommend}: Food recommendations ("NIN_recommendations.csv")
}}
}
\value{
NIN Lancet data as MAgPIE object at ISO country level
}
\description{
Convert data from the NIN Lancet Comission to ISO country level.
}
\examples{

\dontrun{
a <- readSource(type="NIN",subtype="cons_data")
}
}
\seealso{
\code{\link[madrat]{readSource}}
}
\author{
Isabelle Weindl
}
