% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTimberShare.R
\name{convertTimberShare}
\alias{convertTimberShare}
\title{Converts timber share
Update dd-Jmm-jjjj - Please add comment if changes made here (Abhi)}
\usage{
convertTimberShare(x)
}
\arguments{
\item{x}{MAgPIE object to be converted}
}
\value{
A MAgPIE object containing country disaggregated data
}
\description{
Converts timber share
Update dd-Jmm-jjjj - Please add comment if changes made here (Abhi)
}
\examples{
\dontrun{
a <- readSource("TimberShare", convert = FALSE)
}
}
\author{
Abhijeet Mishra
}
