% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct2ndBioDem.R
\name{correct2ndBioDem}
\alias{correct2ndBioDem}
\title{correct2ndBioDem}
\usage{
correct2ndBioDem(x)
}
\arguments{
\item{x}{magclass object provided by the read function}
}
\value{
magclass object
}
\description{
apply lowpass filter on 2ndBioDem to filter out high frequency data
}
\seealso{
\code{\link{calc2ndBioDem}}
}
\author{
Florian Humpenoeder
}
