% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctBrennan2022.R
\name{correctBrennan2022}
\alias{correctBrennan2022}
\title{correctBrennan2022}
\usage{
correctBrennan2022(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
magpie object on cellular level
}
\description{
correct data for Critical Connectivity
Areas (Brennan et al. 2022).
}
\examples{
\dontrun{
readSource("Brennan2022", convert = "onlycorrect")
}

}
\seealso{
\code{\link{readBrennan2022}}
}
\author{
Patrick v. Jeetze
}
