% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctCopernicus.R
\name{correctCopernicus}
\alias{correctCopernicus}
\title{correctCopernicus}
\usage{
correctCopernicus(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
magpie object on cellular level
}
\description{
correct Copernicus data.
}
\examples{
\dontrun{
readSource("Copernicus", convert = "onlycorrect")
}

}
\seealso{
\code{\link{readCopernicus}}
}
\author{
Patrick v. Jeetze
}
