% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctDinerstein2020.R
\name{correctDinerstein2020}
\alias{correctDinerstein2020}
\title{correctDinerstein2020}
\usage{
correctDinerstein2020(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
magpie object on cellular level
}
\description{
correct data for the Global Safety Net conservation priority
areas (Dinerstein et al. 2020).
}
\examples{
\dontrun{
readSource("Dinerstein2020", convert = "onlycorrect")
}

}
\seealso{
\code{\link{readDinerstein2020}}
}
\author{
Patrick v. Jeetze
}
