% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctForestGrassi2023.R
\name{correctForestGrassi2023}
\alias{correctForestGrassi2023}
\title{correctForestGrassi2023}
\usage{
correctForestGrassi2023(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
magpie object on cellular level
}
\description{
Sets cells without data in the forest map of Grassi et al. (2023) to zero.
}
\examples{
\dontrun{
readSource("ForestGrassi2023", convert = "onlycorrect")
}
}
\seealso{
\code{\link{readForestGrassi2023}}
}
\author{
Florian Humpenöder
}
