% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctGLW3all.R
\name{correctGLW3all}
\alias{correctGLW3all}
\title{correctGLW3all}
\usage{
correctGLW3all(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
Magpie objects with results on cellular level, weight, unit and description.
}
\description{
Correct GLW3all data
}
\examples{

\dontrun{
  readSource("GLW3all", subtype = "Da_Ct_2010", convert = "onlycorrect")
}

}
\seealso{
\code{\link{readGLW3all}}
}
\author{
Bin Lin
}
