% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctGLW4.R
\name{correctGLW4}
\alias{correctGLW4}
\title{correctGLW4}
\usage{
correctGLW4(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
Magpie object with NA and negative values replaced by 0.
}
\description{
Replaces NA values and any negative artefacts with zero in
  GLW 4 gridded livestock rasters (reference years 2015 and 2020).
}
\examples{
\dontrun{
  readSource("GLW4", subtype = "Da_Ct_2015", convert = "onlycorrect")
}
}
\seealso{
\code{\link{readGLW4}}
}
\author{
Bin Lin
}
