% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctKeyBiodiversityAreas.R
\name{correctKeyBiodiversityAreas}
\alias{correctKeyBiodiversityAreas}
\title{correctKeyBiodiversityAreas}
\usage{
correctKeyBiodiversityAreas(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
magpie object on cellular level
}
\description{
correct data for Key Biodiversity Areas.
}
\examples{
\dontrun{
readSource("KeyBiodiversityAreas", convert = "onlycorrect")
}

}
\seealso{
\code{\link{readKeyBiodiversityAreas}}
}
\author{
Patrick v. Jeetze
}
