% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctLandMark.R
\name{correctLandMark}
\alias{correctLandMark}
\title{correctLandMark}
\usage{
correctLandMark(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
magpie object on cellular level
}
\description{
correct data of the LandMark IPLC lands.
}
\examples{
\dontrun{
readSource("LandMark", convert = "onlycorrect")
}

}
\seealso{
\code{\link{readLandMark}}
}
\author{
Patrick v. Jeetze
}
