% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctLUH2UrbanFuture.R
\name{correctLUH2UrbanFuture}
\alias{correctLUH2UrbanFuture}
\title{correctLUH2UrbanFuture}
\usage{
correctLUH2UrbanFuture(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
magpie object on cellular level
}
\description{
correct LUH2v2 urban future data
}
\examples{
\dontrun{
readSource("LUH2UrbanFuture", convert = "onlycorrect")
}
}
\seealso{
\code{\link{readLUH2UrbanFuture}}
}
\author{
Patrick v. Jeetze
}
