% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctNoon2022.R
\name{correctNoon2022}
\alias{correctNoon2022}
\title{correctNoon2022}
\usage{
correctNoon2022(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
magpie object on cellular level
}
\description{
correct irrecoverable carbon data from Noon et al. (2022).
}
\examples{
\dontrun{
readSource("Noon2022", convert = "onlycorrect")
}

}
\seealso{
\code{\link{readNoon2022}}
}
\author{
Patrick v. Jeetze
}
