% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctOzoneYieldShock.R
\name{correctOzoneYieldShock}
\alias{correctOzoneYieldShock}
\title{correctOzoneYieldShock}
\usage{
correctOzoneYieldShock(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
x corrected magpie object containing all ISO countries
}
\description{
correct Ozone Yield shock data
}
\examples{
\dontrun{
readSource("OzoneShock", convert="onlycorrect")
}
}
\author{
Jake Tommey
}
