% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctProtectedAreaBaseline.R
\name{correctProtectedAreaBaseline}
\alias{correctProtectedAreaBaseline}
\title{correctProtectedAreaBaseline}
\usage{
correctProtectedAreaBaseline(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
magpie object on cellular level
}
\description{
correct protected area baseline data
}
\examples{
\dontrun{
readSource("ProtectedAreaBaseline", convert = "onlycorrect")
}

}
\seealso{
\code{\link{readProtectedAreaBaseline}}
}
\author{
Patrick v. Jeetze
}
