% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctS4Nproject_input.R
\name{correctS4Nproject_input}
\alias{correctS4Nproject_input}
\title{correctS4Nproject_input}
\usage{
correctS4Nproject_input(x)
}
\arguments{
\item{x}{magpie object}
}
\value{
magpie object at country-level resolution
}
\description{
corrects IMAGE inputs of total bioenergy (1st gen, 2nd gen and residues) demand and co2 prices
}
\examples{

\dontrun{ a <- readSource("S4Nproject_input", aggregate=FALSE)
}
}
\seealso{
\code{\link[madrat]{readSource}}
}
\author{
Felicitas Beier
}
