% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadForestGrassi2023.R
\name{downloadForestGrassi2023}
\alias{downloadForestGrassi2023}
\title{downloadForestGrassi2023}
\usage{
downloadForestGrassi2023()
}
\description{
Downloads the 0.5 degree map of intact and non-intact forest area of Grassi et al. (2023)
from their background data on Zenodo.
}
\examples{
\dontrun{
downloadSource("ForestGrassi2023")
}
}
\seealso{
\code{\link{readForestGrassi2023}}
}
\author{
Florian Humpenöder
}
