% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadGFWLossByDriver.R
\name{downloadGFWLossByDriver}
\alias{downloadGFWLossByDriver}
\title{downloadGFWLossByDriver}
\usage{
downloadGFWLossByDriver()
}
\description{
Downloads tree cover loss by country, year and dominant driver from the Global
Forest Watch data API, table \code{gadm__tcl__iso_change}: the UMD/Hansen annual tree cover loss
cross-tabulated with the 1 km dominant-driver map of Sims et al. (2025).
}
\examples{
\dontrun{
madrat::downloadSource("GFWLossByDriver")
}
}
\seealso{
\code{\link{readGFWLossByDriver}}
}
\author{
Michael Crawford
}
