% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadH08evapotranspiration.R
\name{downloadH08vapotranspiration}
\alias{downloadH08vapotranspiration}
\title{downloadH08evapotranspiration}
\usage{
downloadH08vapotranspiration(subtype = "H08:mri-esm2-0:historical")
}
\arguments{
\item{subtype}{Switch between different inputs}
}
\description{
Download water models evapotranspiration data
}
\examples{

\dontrun{readSource("H08evapotranspiration",  convert="onlycorrect")}
}
\author{
Marcos Alves
}
