% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadSPAM.R
\name{downloadSPAM}
\alias{downloadSPAM}
\title{downloadSPAM}
\usage{
downloadSPAM(subtype)
}
\arguments{
\item{subtype}{Type of SPAM data to be downloaded. Available are "harvestedArea" and "physicalArea".}
}
\description{
download SPAM 2010 v2.0 Global Data
}
\author{
David Hoetten
}
