% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrland-package.R
\docType{package}
\name{mrland-package}
\alias{mrland-package}
\alias{mrland}
\title{MadRaT land data package}
\description{
The package provides land related data via the madrat framework.
}
\author{
Maintainer: Jan Philipp Dietrich <dietrich@pik-potsdam.de>
}
