% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAQUASTAT.R
\name{readAQUASTAT}
\alias{readAQUASTAT}
\title{readAQUASTAT}
\usage{
readAQUASTAT(subtype = "ConsAgri")
}
\arguments{
\item{subtype}{\itemize{
\item \code{ConsAgri}:      4454|Conservation agriculture area (1000 ha)
                            4454_conservation_agriculture_area_in_1000_ha.csv
\item \code{ConsAgriShare}: 4455|Commoditiy Balance LivestockConservation
                                 agriculture area as % of arable land area (%)
                            4455_conservation_agriculture_area_as_share_of_
                                arable_land_areas.csv)
\item \code{rf2irRatio}:    Ratio between rainfed and irrigated yields (%)
                            Ratio_between_rainfed_and_irrigated_yields.csv
}}
}
\value{
magpie objects with results on contury level
}
\description{
Read in data based on AQUASTAT database
             (https://www.fao.org/aquastat/statistics/query/index.html)
}
\examples{
\dontrun{
readSource("AQUASTAT", subtype = "ConsAgri", convert = TRUE)
}

}
\author{
Kristine Karstens
}
