% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDinerstein2020.R
\name{readDinerstein2020}
\alias{readDinerstein2020}
\title{readDinerstein2020}
\usage{
readDinerstein2020(subtype = "GSN:distinct_species_assemblages")
}
\arguments{
\item{subtype}{Defines which cluster (see Dinerstein et al. 2020) of the
Global Safety Net is returned.The different subtypes for land are:
\code{"GSN:distinct_species_assemblages"}, \code{"GSN:rare_phenomena"},
\code{"GSN:areas_of_intactness"}, \code{"GSN:climate_stabilisation_tier1"} and
\code{"GSN:climate_stabilisation_tier2"}.}
}
\value{
Returns magpie objects with the land area covered by the Global Safety Net that
is NOT already covered by Key Biodiversity Areas and was unprotected in 2020.
}
\description{
Reads Global Safety Net data set published by Dinerstein, E., Joshi,
A. R., Vynne, C., Lee, A. T. L., Pharand-Deschênes, F., França, M., Fernando, S.,
Birch, T., Burkart, K., Asner, G. P., & Olson, D. (2020). A “Global Safety Net”
to reverse biodiversity loss and stabilize Earth’s climate. Science Advances, 6(36),
eabb2824. https://doi.org/10.1126/sciadv.abb2824

Protected areas and Key Biodiversity Areas were masked at
a spatial resolution of 10 arc seconds before aggregating the data to 0.5°.
}
\examples{
\dontrun{
readSource("Dinerstein2020", convert = "onlycorrect")
}

}
\author{
Patrick v. Jeetze
}
