% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEATLancet.R
\name{readEATLancet}
\alias{readEATLancet}
\title{readEATLancet}
\usage{
readEATLancet(subtype)
}
\arguments{
\item{subtype}{Type of EAT-Lancet data that should be read. Available types are:
\itemize{
\item \code{cons_data}: Consumption analysis ("EAT_Lancet_cons_data.csv")
\item \code{recommend}: Food recommendations ("EAT_Lancet_recommendations.csv")
}}
}
\value{
magpie object containing EAT-Lancet Comission data
}
\description{
Read in data from the EAT-Lancet Commission

Read in data from:
Food in the Anthropocene: the EAT-Lancet Commission on healthy diets from sustainable food systems, Lancet 2019
https://doi.org/10.1016/S0140-6736(18)31788-4
}
\examples{
\dontrun{
a <- readSource(type = "EATLancet", subtype = "cons_data")
}

}
\seealso{
\code{\link[madrat]{readSource}}
}
\author{
Isabelle Weindl, Jan Philipp Dietrich, Felicitas Beier
}
