% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readForestryProductionRatio.R
\name{readForestryProductionRatio}
\alias{readForestryProductionRatio}
\title{Read Forestry Production Ratio}
\usage{
readForestryProductionRatio()
}
\value{
magpie object of the proportion of production coming from plantations
}
\description{
Read Forestry Production Ratio
}
\examples{

\dontrun{ a <- readSource("ForestryProductionRatio")
}

}
\seealso{
\code{\link[madrat]{readSource}}
}
\author{
Abhijeet Mishra
}
