% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGFWLossByDriver.R
\name{readGFWLossByDriver}
\alias{readGFWLossByDriver}
\title{readGFWLossByDriver}
\usage{
readGFWLossByDriver()
}
\value{
magpie object, ISO country x year (2001-2025) x driver class, Mha
}
\description{
Reads the Global Forest Watch extract downloaded by
\code{\link{downloadGFWLossByDriver}}: tree cover loss by country, year and driver, in Mha.
}
\examples{
\dontrun{
a <- readSource("GFWLossByDriver")
}
}
\seealso{
\code{\link{downloadGFWLossByDriver}}, \code{\link{convertGFWLossByDriver}}
}
\author{
Michael Crawford
}
