% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readH08evapotranspiration.R
\name{readH08evapotranspiration}
\alias{readH08evapotranspiration}
\title{readH08evapotranspiration}
\usage{
readH08evapotranspiration(subtype = "H08:mri-esm2-0:historical")
}
\arguments{
\item{subtype}{Switch between different inputs}
}
\value{
MAgPIE objects with results on cellular level.
}
\description{
Read evapotranspiration data
}
\examples{
\dontrun{
readSource("H08evapotranspiration", subtype, convert = "onlycorrect")
}

}
\seealso{
\code{\link{readH08evapotranspiration}}
}
\author{
Marcos Alves
}
