% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readInocencio2007.R
\name{readInocencio2007}
\alias{readInocencio2007}
\title{readInocencio2007}
\usage{
readInocencio2007()
}
\value{
magpie object of unit costs of irrigation projects
}
\description{
reads in average unit cost of irrigation projects per region and
             by purpose of project from Inocencio et al. (2007) IWMI Research Report 109
}
\examples{
\dontrun{ a <- readSource("Inocencio2007")
}

}
\seealso{
\code{\link[madrat]{readSource}}
}
\author{
Felicitas Beier
}
