% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readKeyBiodiversityAreas.R
\name{readKeyBiodiversityAreas}
\alias{readKeyBiodiversityAreas}
\title{readKeyBiodiversityAreas}
\usage{
readKeyBiodiversityAreas(subtype = "unprotected")
}
\arguments{
\item{subtype}{"unprotected" or "all"}
}
\value{
Returns magpie objects with the area covered by unprotected Key Biodiversity Areas per grid cell
}
\description{
Reads land area covered by for Key Biodiversity Areas
(https://www.keybiodiversityareas.org/) that was unprotected in 2020.
Protected areas were masked at a spatial resolution of 10 arc seconds
before aggregating the data to 0.5°.
}
\examples{
\dontrun{
readSource("KeyBiodiversityAreas", convert = "onlycorrect")
}

}
\author{
Patrick v. Jeetze
}
