% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLandMark.R
\name{readLandMark}
\alias{readLandMark}
\title{readLandMark}
\usage{
readLandMark(subtype = "delineated")
}
\arguments{
\item{subtype}{Whether \code{"delineated"} or \code{"indicative"} lands and
territories are returned.}
}
\value{
Returns magpie objects with the land area covered by the IPLC lands
that was unprotected in 2020.
}
\description{
Reads Indigenous Peoples’ and Local Community Lands from LandMark
"LandMark. 2025. Indicative Areas of Indigenous and Community Land Rights. Data
file from LandMark: The Global Platform of Indigenous and Community Lands.
Available at www.landmarkmap.org."

Protected areas (WDPA+China) were masked at a spatial resolution of
10 arcseconds before aggregating the data to 0.5°.
}
\examples{
\dontrun{
readSource("LandMark", convert = "onlycorrect")
}

}
\author{
Patrick v. Jeetze
}
