% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNIN.R
\name{readNIN}
\alias{readNIN}
\title{Read in data from the NIN-Lancet Comission}
\usage{
readNIN(subtype)
}
\arguments{
\item{subtype}{Type of NIN-Lancet data that should be read. Available types are:
\itemize{
\item \code{cons_data}: Consumption analysis ("NIN_Lancet_cons_data.csv")
\item \code{recommend}: Food recommendations ("NIN_Lancet_recommendations.csv")
}}
}
\value{
magpie object containing NIN-Lancet Comission data
}
\description{
Read in data from:
Food in the Anthropocene: the NIN-Lancet Commission on healthy diets from sustainable food systems, Lancet 2019
https://doi.org/10.1016/S0140-6736(18)31788-4
}
\examples{

\dontrun{ a <- readSource(type="NIN",subtype="cons_data")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Isabelle Weindl, Jan Philipp Dietrich
}
