% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readOzoneYieldShock.R
\name{readOzoneYieldShock}
\alias{readOzoneYieldShock}
\title{readOzoneYieldShock}
\usage{
readOzoneYieldShock()
}
\value{
MAgPIE object with country level yield shock data for year 2050.
}
\description{
read Ozone Yield Shock
Data from the EAT-Lancet deepdive on Ozone shock effects on crop yields.
}
\examples{
\dontrun{
readSource("OzoneShock", convert = "onlycorrect")
}
}
\author{
Jake Tommey
}
