% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readProtectedAreaBaseline.R
\name{readProtectedAreaBaseline}
\alias{readProtectedAreaBaseline}
\title{readProtectedAreaBaseline}
\usage{
readProtectedAreaBaseline()
}
\value{
Returns magpie object with the protected area separated for each land type
        (cropland, pasture, forest, other land) per grid cell from 1995 to 2020.
}
\description{
Reads spatial land cover information within protected areas.
             Land cover information for protected areas has been extracted from ESA CCI
             land use/land cover data (https://www.esa-landcover-cci.org/) and data from
             the WDPA data base (https://www.protectedplanet.net).
}
\examples{
\dontrun{
readSource("ProtectedAreaBaseline", convert = "onlycorrect")
}
}
\author{
Patrick v. Jeetze
}
