% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readREMIND.R
\name{readREMIND}
\alias{readREMIND}
\title{readREMIND}
\usage{
readREMIND(subtype)
}
\arguments{
\item{subtype}{A string composed of three items: unit, revision and indicator.
Unit can be either "intensive" or "extensive", revision is the input data
revision, and indicator is the name of the REMIND indicator}
}
\value{
MAgPIE object with regional aggregation of REMIND H12
}
\description{
Reads in a reporting mif file from REMIND
}
\examples{

\dontrun{
readSource("REMIND",aggregate=FALSE)
}
}
\seealso{
\code{\link[madrat]{readSource}}
}
\author{
David Klein
}
