% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readS4Nproject_input.R
\name{readS4Nproject_input}
\alias{readS4Nproject_input}
\title{readS4Nproject_input}
\usage{
readS4Nproject_input(subtype = "co2prices")
}
\arguments{
\item{subtype}{IMAGE input to be read in: co2prices or bioenergy}
}
\value{
magpie object at country-level resolution
}
\description{
reads in total bioenergy (1st gen, 2nd gen and residues)
demand and co2 prices from IMAGE model for Sim4Nexus project
}
\examples{
\dontrun{
a <- readSource("S4Nproject_input", convert = "onlycorrect", aggregate = FALSE)
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Felicitas Beier
}
