% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSPAM.R
\name{readSPAM}
\alias{readSPAM}
\title{readSPAM}
\usage{
readSPAM(subtype)
}
\arguments{
\item{subtype}{Type of SPAM data to be readed. Available are "harvestedArea" and "physicalArea".}
}
\value{
magpie object containing output of the toolbox
}
\description{
Reads SPAM 2010 v2.0 Global Data
The items in the technology dimension have the following significations:
A =	all technologies together, ie complete crop; I =	irrigated portion of crop;
H =	rainfed high inputs portion of crop; L =	rainfed low inputs portion of crop;
S =	rainfed subsistence portion of crop; R =	rainfed portion of crop (= A - I, or H + L + S)
All data is in ha.
}
\examples{
\dontrun{
A <- readSource("SPAM", subtype = "harvestedArea", aggregate = FALSE)
}

}
\seealso{
\code{\link{readSource}}
}
\author{
David Hoetten
}
