% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readStrefler2021.R
\name{readStrefler2021}
\alias{readStrefler2021}
\title{readStrefler2021}
\usage{
readStrefler2021(subtype)
}
\arguments{
\item{subtype}{Either "intensive" or "extensive"}
}
\value{
MAgPIE object with regional aggregation of REMIND H12
}
\description{
Reads in a reporting mif file from REMIND
}
\examples{

\dontrun{ 
readSource("Strefler2021",aggregate=FALSE)
}
}
\seealso{
\code{\link[madrat]{readSource}}
}
\author{
Florian Humpenöder
}
