% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTimberShare.R
\name{readTimberShare}
\alias{readTimberShare}
\title{Read Share of timber predicted to come from plantations based on FAO Brown study}
\usage{
readTimberShare(subtype = "abare")
}
\arguments{
\item{subtype}{Data subtype available is abare and brown}
}
\value{
magpie object of the proportion of production coming from plantations
}
\description{
Read Share of timber predicted to come from plantations based on FAO Brown study
}
\examples{

\dontrun{
a <- readSource("TimberShare")
}

}
\seealso{
\code{\link[madrat]{readSource}}
}
\author{
Abhijeet Mishra
}
