% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWBirrigation.R
\name{readWBirrigation}
\alias{readWBirrigation}
\title{readWBirrigation}
\usage{
readWBirrigation()
}
\value{
magpie object of the WBirrigation data
}
\description{
reads in World bank irrigation data: WBirrigation data .csv file as magclass object
from Jones, William I. 1995. "World Bank and Irrigation." Washington, D.C.: World Bank. 
Bonsch et al. (2015) "Environmental Flow Provision: Implications for Agricultural Water and Land-Use at the Global Scale": Table A1 - Investment costs for expanding irrigation infrastructure in US$ per hectare.
Based on: World Bank Irrigation Investment Cost Data. 
William I. Jones(1991) "The World Bank and Irrigation" (World Bank Operations Evaluation Study)
}
\examples{

\dontrun{ a <- readSource(type="WBirrigation")
}

}
\seealso{
\code{\link[madrat]{readSource}}
}
\author{
Lavinia Baumstark
}
