% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolImportSupplyRatioAggregate.R
\name{toolImportSupplyRatioAggregate}
\alias{toolImportSupplyRatioAggregate}
\title{toolImportSupplyRatioAggregate
This function aggregates bilateral import supply ratios (amount exported from
country X divided by domestic supply of country Y) from country to regional
level, using two different weights for bilateral spatial dimensions
1.1 and 1.2, where dimension 1.1 is the exporters (no weight, simple summation),
while dimension 1.2 are the importers, weighted by domestic supply.
THIS ORDER IS IMPORTANT.}
\usage{
toolImportSupplyRatioAggregate(x, rel, weight)
}
\arguments{
\item{x}{magclass object that should be aggregated}

\item{rel}{relation matrix containing a region mapping.}

\item{weight}{weight to use for dimension 1.2}
}
\value{
return: returns region aggregated import supply ratios
}
\description{
toolImportSupplyRatioAggregate
This function aggregates bilateral import supply ratios (amount exported from
country X divided by domestic supply of country Y) from country to regional
level, using two different weights for bilateral spatial dimensions
1.1 and 1.2, where dimension 1.1 is the exporters (no weight, simple summation),
while dimension 1.2 are the importers, weighted by domestic supply.
THIS ORDER IS IMPORTANT.
}
\author{
David M Chen
}
