% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolThornthwaite.R
\name{toolThornthwaite}
\alias{toolThornthwaite}
\title{toolThornthwaite}
\usage{
toolThornthwaite(temp, lat)
}
\arguments{
\item{temp}{monthly mean temperature}

\item{lat}{lattitude of the cell}
}
\value{
PET in magclass format
}
\description{
This tool to calculate potential evapotranspiration (PET) based on the
             Thornthwaite function (see SPEI::thornthwaite)
}
\author{
Kristine Karstens
}
