% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAtmosphericDepositionRates.R
\name{calcAtmosphericDepositionRates}
\alias{calcAtmosphericDepositionRates}
\title{calcAtmosphericDepositionRates}
\usage{
calcAtmosphericDepositionRates(cellular = FALSE, cells = "lpjcell")
}
\arguments{
\item{cellular}{TRUE for results on 0.5 degree grid.}

\item{cells}{magpiecell (59199 cells) or lpjcell (67420 cells)}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Conputes Atmospheric (nitrogen) deposition rates per area on different land-use types.
}
\examples{
\dontrun{
calcOutput("AtmosphericDepositionRates")
}

}
\seealso{
\code{\link[mrcommons]{calcAtmosphericDeposition}},
\code{\link[mrcommons]{calcNitrogenBudgetCropland}}
}
\author{
Benjamin Leon Bodirsky
}
