% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcForestFireLoss.R
\name{calcForestFireLoss}
\alias{calcForestFireLoss}
\title{calcForestFireLoss}
\usage{
calcForestFireLoss()
}
\value{
MAgPIE object with FRA 2020 forest fire area loss
}
\description{
Calculate how much loss of forest area happens due to fire disturbances based on FRA 2020 data
}
\examples{

\dontrun{ 
calcOutput("ForestFireLoss",aggregate=FALSE)
}


}
\seealso{
\code{\link[mrfaocore]{readFRA2020}}
}
\author{
Abhijeet Mishra
}
