% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcForestLossShare.R
\name{calcForestLossShare}
\alias{calcForestLossShare}
\title{calcForestLossShare}
\usage{
calcForestLossShare()
}
\value{
MAgPIE object with share of area lost in forests due to different drivers
}
\description{
Calculates which share of forest land is lost due to different drivers
}
\examples{

\dontrun{ 
calcOutput("ForestLossShare",aggregate=FALSE)
}


}
\seealso{
\code{\link{readForestLossDrivers}}
}
\author{
Abhijeet Mishra
}
