% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHalfEarth.R
\name{calcHalfEarth}
\alias{calcHalfEarth}
\title{calcHalfEarth}
\usage{
calcHalfEarth(cells = "lpjcell", nclasses = "seven")
}
\arguments{
\item{cells}{(deprecated) always lpjcell (67420 cells)}

\item{nclasses}{Options are either "seven" or "nine".
\itemize{
\item "seven" separates primary and secondary forest and includes "crop", "past", "forestry",
"primforest", "secdforest", "urban" and "other"
\item "nine" adds the separation of pasture and rangelands, as well as a differentiation of
primary and secondary non-forest vegetation and therefore returns "crop", "past", "range",
"forestry", "primforest", "secdforest", "urban", "primother" and "secdother"
}}
}
\value{
magpie object in cellular resolution with different protection options in conservation priority areas
}
\description{
Function calculates land area in 'Half Earth' conservation priority area
}
\examples{
\dontrun{
calcOutput("HalfEarth", aggregate = FALSE)
}

}
\author{
Patrick v. Jeetze, Felicitas Beier
}
