% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrecoverableCarbonLand.R
\name{calcIrrecoverableCarbonLand}
\alias{calcIrrecoverableCarbonLand}
\title{calcIrrecoverableCarbonLand}
\usage{
calcIrrecoverableCarbonLand(
  maginput = TRUE,
  nclasses = "seven",
  cells = "lpjcell"
)
}
\arguments{
\item{maginput}{Whether data should be transformed (based on luh3 data) to match land use types used in MAgPIE.}

\item{nclasses}{If \code{magpie_input = TRUE}. Options are either "seven" or "nine". Note that by default,
the protected area is reported for urban land and forestry is zero.
\itemize{
\item "seven" separates primary and secondary forest and includes
"crop", "past", "forestry", "primforest", "secdforest", "urban" and "other"
\item "nine" adds the separation of pasture and rangelands, as well as a
differentiation of primary and secondary non-forest vegetation and therefore returns
"crop", "past", "range", "forestry", "primforest", "secdforest", "urban", "primother" and "secdother"
}}

\item{cells}{magpiecell (59199 cells) or lpjcell (67420 cells)}
}
\value{
List with a magpie object
}
\description{
Returns unprotected land area (Mha) that covers 50 %, 75% and
99% of irrecoverable carbon as defined in Noon et al (2022).
}
\examples{
\dontrun{
calcOutput("calcIrrecoverableCarbonLand", aggregate = FALSE)
}

}
\seealso{
\code{\link{readNoon2022}}
}
\author{
Patrick v. Jeetze
}
