% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigUnitCost.R
\name{calcIrrigUnitCost}
\alias{calcIrrigUnitCost}
\title{calcIrrigUnitCost}
\usage{
calcIrrigUnitCost()
}
\value{
MAgPIE object
}
\description{
This function returns unit costs for irrigation expansion (new_totalCost)
             and maintenance or rehabilitation (rehab_hardwareCost)
}
\examples{
\dontrun{
calcOutput("calcIrrigUnitCost")
}

}
\seealso{
\code{\link[madrat]{calcOutput}}, \code{\link{readInocencio2007}},
\code{\link{convertInocencio2007}}
}
\author{
Felicitas Beier
}
