% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLPJmLClimateInput.R
\name{calcLPJmLClimateInput}
\alias{calcLPJmLClimateInput}
\title{calcLPJmLClimateInput}
\usage{
calcLPJmLClimateInput(
  climatetype = "IPSL-CM6A-LR:ssp126",
  variable = "temperature:annual_mean",
  stage = "harmonized2020",
  lpjmlVersion = "LPJmL4_for_MAgPIE_44ac93de"
)
}
\arguments{
\item{climatetype}{Switch between different climate scenario}

\item{variable}{Switch between different climate inputs and temporal resolution}

\item{stage}{Degree of processing: raw, smoothed, harmonized, harmonized2020}

\item{lpjmlVersion}{LPJmL Version hand over}
}
\value{
magpie object in cellular resolution
}
\description{
Handle LPJmL climate input data and its time behaviour
             (smoothing and harmonizing to baseline)
}
\examples{
\dontrun{
calcOutput("LPJmLClimateInput",
           climatetype = "IPSL-CM6A-LR:ssp126",
           variable = "temperature:annual_mean")
}

}
\author{
Marcos Alves, Kristine Karstens, Felicitas Beier
}
