% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMulticroppingYieldIncrease.R
\name{calcMulticroppingYieldIncrease}
\alias{calcMulticroppingYieldIncrease}
\title{calcMulticroppingYieldIncrease}
\usage{
calcMulticroppingYieldIncrease(
  selectyears,
  lpjml,
  climatetype,
  fallowFactor = 0.75,
  minThreshold = 100,
  areaMask = "potential:endogenous",
  crops = "standard"
)
}
\arguments{
\item{selectyears}{Years to be returned}

\item{lpjml}{LPJmL version required for respective inputs as single string: "crop" version}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}

\item{fallowFactor}{Factor determining yield reduction in off season due to
fallow period between harvest of first (main) season and
sowing of second (off) season}

\item{minThreshold}{Minimum threshold of grass GPP in crop growing period
and crop yield to exclude low yielding cells
Unit of the threshold is gC/m^2.
Default: 100 gC/m^2}

\item{areaMask}{Multicropping area mask to be used
"none": no mask applied (only for development purposes)
"actual:total": currently multicropped areas calculated from total harvested areas
                and total physical areas per cell from readLanduseToolbox
"actual:crop" (crop-specific), "actual:irrigation" (irrigation-specific),
"actual:irrig_crop" (crop- and irrigation-specific)
"potential:endogenous": potentially multicropped areas given
                        temperature and productivity limits
"potential:exogenous": potentially multicropped areas given
                       GAEZ suitability classification}

\item{crops}{standard: default crops,
proxy: proxy crops for LPJmL to MAgPIE mapping and treatment of perennials}
}
\value{
magpie object in cellular resolution
}
\description{
Calculates yield increase achieved through multiple cropping
             (as factor of off season to main season crop yield) under
             irrigated and rainfed conditions respectively.
             Optionally: return which grid cells are potentially suitable for
             multiple cropping activities under rainfed and irrigated conditions.
             Calculation is based on grassland gross primary production (GPP)
             in the growing period of the respective crop and annual grass GPP.
}
\examples{
\dontrun{
calcOutput("MulticroppingYieldIncrease", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
