% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNitrogenFixationNdfa.R
\name{calcNitrogenFixationNdfa}
\alias{calcNitrogenFixationNdfa}
\title{calcNitrogenFixationNdfa}
\usage{
calcNitrogenFixationNdfa()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates the share of N in biomass derived from biological fixation
}
\examples{
\dontrun{
calcOutput("NitrogenFixationNdfa")
}

}
\author{
Benjamin Leon Bodirsky
}
