% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPastureYield.R
\name{calcPastureYield}
\alias{calcPastureYield}
\title{calcPastureYield}
\usage{
calcPastureYield(range_pastr = FALSE)
}
\arguments{
\item{range_pastr}{Boolean value indicating if the grass yields
should be split between rangelands and pastures.}
}
\value{
Pasture yields and corresponding weights as a list of two MAgPIE objects
}
\description{
Provides pasture yields defined as
             ratio of grazed biomass to grazed area
}
\examples{
\dontrun{
calcOutput("PastureYield")
}
}
\seealso{
\code{\link{calcOutput}}, \code{\link{calcFAOmassbalance}},
\code{\link{readSource}}
}
\author{
Isabelle Weindl, Marcos Alves
}
