% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSNVTargetCropland.R
\name{calcSNVTargetCropland}
\alias{calcSNVTargetCropland}
\title{calcSNVTargetCropland}
\usage{
calcSNVTargetCropland(maginput = TRUE)
}
\arguments{
\item{maginput}{Whether data should be corrected to align with cropland
initialised in MAgPIE.}
}
\value{
List with a magpie object
}
\description{
Returns cropland area (Mha) that requires relocation in response of
maintaining 20% or 50% semi-natural vegetation in farmed landscapes.
}
\examples{
\dontrun{
calcOutput("SNVTargetCropland", aggregate = FALSE)
}

}
\seealso{
\code{\link{readCopernicus}}
}
\author{
Patrick v. Jeetze
}
