% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTradeBilateralBalanceFlow.R
\name{calcTradeBilateralBalanceFlow}
\alias{calcTradeBilateralBalanceFlow}
\title{calcTradeBilateralBalanceFlow}
\usage{
calcTradeBilateralBalanceFlow(balanceflow = "trade")
}
\arguments{
\item{balanceflow}{"trade" or "total"}
}
\value{
Self import to dupply ratio
}
\description{
Calculates balanceflows for trade and total mass balance
to match historical production
}
\examples{
\dontrun{
a <- calcTrademportSupplyStdDev()
}
}
\seealso{
\code{\link[madrat]{calcOutput}}, \code{\link[mrcommons]{calcFAOmassbalance}}
}
\author{
David M Chen
}
