% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcYieldsCalibrated.R
\name{calcYieldsCalibrated}
\alias{calcYieldsCalibrated}
\title{calcYieldsCalibrated}
\usage{
calcYieldsCalibrated(
  source = c(lpjml = "ggcmi_phase3_nchecks_9ca735cb", isimip = NULL),
  climatetype = "GSWP3-W5E5:historical",
  refYear = "y1995",
  cells = "magpiecell",
  multicropping = FALSE
)
}
\arguments{
\item{source}{Defines LPJmL version for main crop inputs and isimip replacement.
For isimip choose crop model/gcm/rcp/co2 combination formatted like this:
"yields:EPIC-IIASA:ukesm1-0-ll:ssp585:default:3b"}

\item{climatetype}{switch between different climate scenarios}

\item{refYear}{reference year for calibration}

\item{cells}{number of cells "magpiecell" for 59199 cells or
"lpjcell" for 67420 cells}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE) and
Multiple Cropping Suitability mask selected
(mask can be:
"none": no mask applied (only for development purposes)
"actual:total": currently multicropped areas calculated from total harvested areas
                and total physical areas per cell from readLanduseToolbox
"actual:crop" (crop-specific), "actual:irrigation" (irrigation-specific),
"actual:cropIrrig" (crop- and irrigation-specific) "total"
"potential:endogenous": potentially multicropped areas given
                        temperature and productivity limits
"potential:exogenous": potentially multicropped areas given
                       GAEZ suitability classification)
(e.g. TRUE:actual:total; TRUE:none; FALSE)}
}
\value{
magpie object in cellular resolution from reference year onwards
}
\description{
This functions calibrates extracted yields from LPJmL to
             FAO country level yields
}
\examples{
\dontrun{
calcOutput("YieldsCalibrated", aggregate = FALSE)
}

}
\author{
Kristine Karstens, Felicitas Beier
}
