% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcYieldsCalibrated.R
\name{calcYieldsCalibrated}
\alias{calcYieldsCalibrated}
\title{calcYieldsCalibrated}
\usage{
calcYieldsCalibrated(
  source = c(lpjml = "ggcmi_phase3_nchecks_9ca735cb", isimip = NULL),
  climatetype = "GSWP3-W5E5:historical",
  refYear = "y1995",
  cells = "magpiecell",
  multicropping = FALSE,
  areaSource = "FAO",
  marginal_land = "magpie"
)
}
\arguments{
\item{source}{Defines LPJmL version for main crop inputs and isimip replacement.
For isimip choose crop model/gcm/rcp/co2 combination formatted like this:
"yields:EPIC-IIASA:ukesm1-0-ll:ssp585:default:3b"}

\item{climatetype}{switch between different climate scenarios}

\item{refYear}{reference year for calibration}

\item{cells}{number of cells "magpiecell" for 59199 cells or
"lpjcell" for 67420 cells}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE) and
Multiple Cropping Suitability mask selected
(mask can be:
"none": no mask applied (only for development purposes)
"actual:total": currently multicropped areas calculated from total harvested areas
                and total physical areas per cell from readLanduseToolbox
"actual:crop" (crop-specific), "actual:irrigation" (irrigation-specific),
"actual:irrig_crop" (crop- and irrigation-specific),
"potential:endogenous": potentially multicropped areas given
                        temperature and productivity limits
"potential:exogenous": potentially multicropped areas given
                       GAEZ suitability classification)
(e.g. TRUE:actual:total; TRUE:none; FALSE)}

\item{areaSource}{data source for croparea used in calculation: FAO or Toolbox
Note: when calibrating multicropped yields, Toolbox croparea should be used}

\item{marginal_land}{Defines which share of marginal land should be included (see options below) and
                     whether suitable land under irrigated conditions ("irrigated"), under rainfed conditions ("rainfed")
                     or suitability under rainfed conditions including currently irrigated land (rainfed_and_irrigated)
                     should be used. Options combined via ":"
                     The different marginal land options are:
\itemize{
\item \code{"all_marginal"}: All marginal land (suitability index between 0-0.33) is included as suitable
\item \code{"q33_marginal"}: The bottom tertile (suitability index below 0.13) of the
marginal land area is excluded.
\item \code{"q50_marginal"}: The bottom  half (suitability index below 0.18) of the
marginal land area is excluded.
\item \code{"q66_marginal"}: The first and second tertile (suitability index below 0.23) of
the marginal land area are excluded.
\item \code{"q75_marginal"}: The first, second and third quartiles (suitability index below 0.25)
of the marginal land are are excluded
\item \code{"no_marginal"}: Areas with a suitability index of 0.33 and lower are excluded.
\item \code{"magpie"}: Returns "all_marginal:rainfed_and_irrigated",
                       "q33_marginal:rainfed_and_irrigated" and
                       "no_marginal:rainfed_and_irrigated" in a magclass object to be used as magpie input.
}}
}
\value{
magpie object in cellular resolution from reference year onwards
}
\description{
This functions calibrates extracted yields from LPJmL to
             FAO country level yields
}
\examples{
\dontrun{
calcOutput("YieldsCalibrated", aggregate = FALSE)
}

}
\author{
Kristine Karstens, Felicitas Beier
}
