% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadInocencio2007.R
\name{downloadInocencio2007}
\alias{downloadInocencio2007}
\title{downloadInocencio2007}
\usage{
downloadInocencio2007()
}
\value{
A list with metadata required by \code{downloadSource()}.
}
\description{
Create (no external download) unit cost table based on manual
             extraction from Inocencio et al. (2007)
}
\seealso{
\code{\link[madrat]{downloadSource}}
}
\author{
Felicitas Beier
}
