% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullMAGPIE.R
\name{fullMAGPIE}
\alias{fullMAGPIE}
\title{fullMAgPIE}
\usage{
fullMAGPIE(rev = numeric_version("0.1"), dev = "")
}
\arguments{
\item{rev}{data revision which should be used as input (numeric_version).}

\item{dev}{For developing purposes, apply changes as per dev flag}
}
\description{
Function that produces the regional data set for running the MAgPIE model.
}
\examples{
\dontrun{
retrieveData("MAGPIE", rev = numeric_version("12"),
             mainfolder = "pathtowhereallfilesarestored")
}

}
\seealso{
\code{\link[madrat]{readSource}}, \code{\link[madrat]{getCalculations}}, \code{\link[madrat]{calcOutput}}
}
\author{
Jan Philipp Dietrich, Benjamin Leon Bodirsky, Florian Humpenoeder, Edna J. Molina Bacca
}
