% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrland-package.R
\docType{package}
\name{mrland-package}
\alias{mrland-package}
\alias{_PACKAGE}
\alias{mrland}
\title{MadRaT land data package}
\description{
The package provides land related data via the madrat framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/mrland}
  \item \doi{10.5281/zenodo.3822083}
  \item Report bugs at \url{https://github.com/pik-piam/mrland/issues}
}

}
\author{
Maintainer: Jan Philipp Dietrich <dietrich@pik-potsdam.de>
}
