% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readForestLossDrivers.R
\name{readForestLossDrivers}
\alias{readForestLossDrivers}
\title{readForestLossDrivers}
\usage{
readForestLossDrivers()
}
\value{
magpie object with the mean annual loss per Curtis region and driver, Mha
}
\description{
Reads Table 1 of Curtis et al. (2018), tree cover loss 2001-2015 by world region
and dominant driver, transcribed by hand from the article into \code{forest_loss_fixed.csv}.
}
\examples{
\dontrun{
a <- readSource("ForestLossDrivers", convert = FALSE)
}
}
\references{
Curtis, P. G., Slay, C. M., Harris, N. L., Tyukavina, A. and Hansen, M. C. (2018)
Classifying drivers of global forest loss. Science 361, 1108-1111. doi:10.1126/science.aau3445
}
\seealso{
\code{\link{calcForestLossByDriver}}
}
\author{
Abhijeet Mishra, Michael Crawford
}
