% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readForestLossDrivers.R
\name{readForestLossDrivers}
\alias{readForestLossDrivers}
\title{Read ForestLossDrivers}
\usage{
readForestLossDrivers()
}
\value{
magpie object of the Curtis et al., 2018 Data
}
\description{
Read-in an Forest loss data (range 2001-2015 but only single annual number her)
(Source:DOI: 10.1126/science.aau3445 Table 1).
}
\examples{
\dontrun{
a <- readSource("ForestLossDrivers")
}

}
\seealso{
\code{\link[madrat]{readSource}}
}
\author{
Abhijeet Mishra
}
