% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGLW3.R
\name{readGLW3}
\alias{readGLW3}
\title{readGLW3}
\usage{
readGLW3(subtype = "Da_Ct_2010")
}
\arguments{
\item{subtype}{Weighting method, livestock species, and reference year
(\code{"<method>_<species>_2010"}):
\itemize{
  \item {Da}: Dasymetric weighting informed by Random Forest
  \item {Aw}: Areal weighting (distributed uniformly in each census unit)
  \itemize{
    \item \code{Ct}: Cattle
    \item \code{Sh}: Sheep
    \item \code{Pg}: Pigs
    \item \code{Bf}: Buffaloes
    \item \code{Ch}: Chickens
    \item \code{Ho}: Horses
    \item \code{Gt}: Goats
    \item \code{Dk}: Ducks
  }
}}
}
\value{
A gridded magpie object with gridded livestock of the world
}
\description{
Reads Gridded Livestock of the World version 3 (GLW 3)
  raster data for reference year 2010, downloaded from Harvard Dataverse.
  Eight livestock species are available with both dasymetric and areal
  weighting.
  Source catalogue:
  https://www.fao.org/livestock-systems/global-distributions/en/
}
\examples{
\dontrun{
readSource("GLW3", subtype = "Da_Ct_2010", convert = FALSE)
}
}
\author{
Marcos Alves, Bin Lin
}
