% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGLW4.R
\name{readGLW4}
\alias{readGLW4}
\title{readGLW4}
\usage{
readGLW4(subtype = "Da_Ct")
}
\arguments{
\item{subtype}{Weighting method and livestock type:
\itemize{
\item {Da}: Dasymetric weighting informed by Random Forest
\item {Aw}: Areal weighting (distributed uniformly in each census)
\itemize{
\item \code{Ch}: Chicken
\item \code{Ct}: Cattle
\item \code{Pg}: Pigs
\item \code{Sh}: Sheep
\item \code{Gt}: Goats
\item \code{Ho}: Horse
\item \code{Dk}: ducks
\item \code{Bf}: Buffaloes
}}}
}
\value{
A gridded magpie object with gridded livstock of the world
}
\description{
reads in Gridded Livestock of the World v4,
downloaded from: https://dataverse.harvard.edu/dataverse/glw_4
}
\author{
David M Chen
}
