% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNIN.R
\name{readNIN}
\alias{readNIN}
\title{Read in data from the NIN recommendations}
\usage{
readNIN(subtype)
}
\arguments{
\item{subtype}{Type of NIN data that should be read. Available types are:
\itemize{
\item \code{cons_data}: Consumption analysis ("NIN_cons_data.csv")
}}
}
\value{
magpie object containing NIN data
}
\description{
Read in data from the NIN recommendations
}
\examples{
\dontrun{
a <- readSource(type = "NIN", subtype = "cons_data")
}

}
\seealso{
\code{\link[madrat]{readSource}}
}
\author{
Isabelle Weindl, Jan Philipp Dietrich
}
