% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNoon2022.R
\name{readNoon2022}
\alias{readNoon2022}
\title{readNoon2022}
\usage{
readNoon2022(subtype = "land:IrrecovCarbon50")
}
\arguments{
\item{subtype}{Defines whether carbon data or land area and related subtypes should be returned (see options below).
Carbon or land subtypes need to be specified via ":"
The different subtypes for land are:
\itemize{
\item \code{"IrrecovCarbon50"}: Unprotected land (specified for different land types) that covers the
highest 50 % of all irrecoverable carbon stocks, as defined in Noon et al. (2022).
\item \code{"IrrecovCarbon75"}: Unprotected land (specified for different land types) that covers the
highest 75 % of all irrecoverable carbon stocks.
\item \code{"IrrecovCarbonAll"}: Unprotected land (specified for different land types) that covers
all irrecoverable carbon stocks.
}}
}
\value{
Returns magpie objects with the area of unprotected irrecoverable carbon land per grid cell
}
\description{
Reads irrecoverble carbon data set published by Noon, M. L., Goldstein, A.,
Ledezma, J. C., Roehrdanz, P. R., Cook-Patton, S. C., Spawn-Lee, S. A., Wright, T. M.,
Gonzalez-Roglich, M., Hole, D. G., Rockström, J., & Turner, W. R. (2022). Mapping the
irrecoverable carbon in Earth’s ecosystems. Nature Sustainability, 5(1),
Article 1. https://doi.org/10.1038/s41893-021-00803-6
}
\examples{
\dontrun{
readSource("Noon2022", convert = "onlycorrect")
}

}
\author{
Patrick v. Jeetze
}
